+++ src/types/vehicle.ts
@@ -1,0 +1,8 @@
+export interface Vehicle {
+  axleWeight: number
+  hasTrailer: false
+  height: number
+  length: number
+  weight: number
+  width: number
+}